+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 - Add SMC-Spoof.kext to avoid triggering `smcupdater`
 - Add Root Volume patching for older machines
   - AppleHDA Patch for 2011 and older (Excluding MacPro4,1+)
-  - AppleBCM5701Ethernet patch for certain 2009-2011 Macs
 - Fix CPU Speed reporting
 - Increment binaries
   - OpenCore c92bcb7 (0.6.8 rolling - 2021-03-20)
@@ -16,6 +15,8 @@
 - Add custom SD Card icon
 - Add automatic codesiging and notarization
 - Fix crashing when CD is present
+- Add custom SSD icon
+- Fix Broadcom Ethernet on older 2009-2011 Macs
 
 ## 0.0.18
 - Disable Vault by default due to breaking installations